feat(ddm): Change endpoint for visibility of metrics
